+++ index.js
@@ -18,6 +18,7 @@ import "./features/alerts/alertOnSpiritMaskUsed";
 import "./features/alerts/alertOnGoldenFish";
 import "./features/alerts/alertOnFishingBagDisabled";
 import "./features/alerts/alertOnHotspotGone";
+import "./features/alerts/alertOnPetLevelUp";
 
 import "./features/overlays/rareCatchesTracker";
 import "./features/overlays/totemTracker";