+++ settings.js
@@ -546,6 +546,14 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
 })
 .addSwitch({
     category: "Alerts",
+    configName: "alertOnPetLevelUp",
+    title: "Alert when a pet reached max level",
+    description: "Shows a title and plays a sound when a pet reached max level.",
+    subcategory: "Pet",
+    value: true
+})
+.addSwitch({
+    category: "Alerts",
     configName: "alertOnSeaCreaturesCountThreshold",
     title: "Alert when sea creatures count hits threshold",
     description: `Shows a title and plays a sound when amount of sea creatures nearby hits the specified threshold. Useful to detect cap when barn fishing. ${RED}Disabled if you have no fishing rod in your hotbar!`,